fix(excel): support excel workbooks that do not have tables
Funny story: it turns out that Excel has worksheets that do not contain tables. These worksheets still have rows and cells and so on, but the container is a range, not a table. A table needs to be created by the user and allows feature such as filtering, sorting, etc. Anyway, this commit adds support for ranges and includes integration tests for workbooks with and workbooks without tables. The error handling code for the more unlikely scenarios (workbooks without worksheets and tables without columns) has been removed, as the testing overhead would exceed the usable value.

fixes #52